requirements.txt
chromadb
openai
PyPDF2
python-docx
sentence-transformers